feat: `BodyRef::without_trailer()` for more obvious access than `*body` or `body.as_ref()` (#198)
